refactoring to biginteger

+++ src/main/java/com/github/fabriciossouza/application/SequenceNumberApplication.java
@@ -3,6 +3,8 @@ package com.github.fabriciossouza.application;
 import com.github.fabriciossouza.domain.SequenceNumberService;
 import jakarta.enterprise.context.ApplicationScoped;
 
+import java.math.BigInteger;
+
 @ApplicationScoped
 public class SequenceNumberApplication {
 
@@ -10,7 +12,7 @@ public class SequenceNumberApplication {
     public SequenceNumberApplication(SequenceNumberService sequenceNumberService) {
         this.sequenceNumberService = sequenceNumberService;
     }
-    public long getSequenceNumber(Integer number) {
+    public BigInteger getSequenceNumber(Integer number) {
         return sequenceNumberService.calculateLabSeqNumber(number);
     }
